Remove EqualsPrim
